requirements.txt
amqp==5.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.14
django-bootstrap-v5==1.0.11
django-crispy-forms==2.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
h11==0.14.0
idna==3.7
kombu==5.3.7
msgpack==1.0.8
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.8
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.3
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
whitenoise==6.7.0
